+++ test/test_languages/testLua.py
@@ -175,3 +175,5 @@ class Test_parser_for_Lua(unittest.TestCase):
         self.assertEqual(2, len(result))
         self.assertEqual('sum', result[0].name)
         self.assertEqual('addn', result[1].name)
+
+