+++ test/test_languages/testSwift.py
@@ -275,3 +275,4 @@ class Test_parser_for_Swift(unittest.TestCase):
         ''')
         self.assertEqual(2, len(result))
 
+